test(sys): add tmq test case
